requirements.txt
beautifulsoup4
awesome-slugify
appdirs
colorama
termcolor
